requirements.txt
annotated-types
anyio
attrs
boto3
botocore
cachetools
certifi
charset-normalizer
cloudpickle
copulas
ctgan
deepecho
distro
Faker
faker
filelock
filetype
fsspec
google-ai-generativelanguage
google-api-core
google-auth
googleapis-common-protos
graphviz
grpcio
grpcio-status
h11
httpcore
httpx
hypothesis
hypothesis-jsonschema
idna
Jinja2
jiter
jmespath
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-core
langchain-google-genai
langchain-openai
langchain-ollama
langchain-text-splitters
langsmith
MarkupSafe
mpmath
mysql-connector-python
narwhals
networkx
numpy
ollama
openai
orjson
packaging
pandas
platformdirs
plotly
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
python-dateutil
python-dotenv
pytz
PyYAML
rdt
referencing
regex
requests
requests-toolbelt
rpds-py
rsa
s3transfer
scikit-learn
scipy
sdmetrics
sdv
setuptools
six
sniffio
sortedcontainers
SQLAlchemy
sympy
tenacity
threadpoolctl
tiktoken
torch
tqdm
typing-inspection
typing_extensions
tzdata
urllib3
zstandard
sqlglot
sqlparse
pymysql